moved ffc msg into own mock module
