+++ test/integration/process-statement-message.test.js
@@ -1,18 +1,8 @@
-const mockSendMessage = jest.fn()
-jest.mock('ffc-messaging', () => {
-  return {
-    MessageSender: jest.fn().mockImplementation(() => {
-      return {
-        sendMessage: mockSendMessage,
-        closeConnection: jest.fn()
-      }
-    })
-  }
-})
-
 const { BlobServiceClient } = require('@azure/storage-blob')
 const config = require('../../app/config/storage')
 const db = require('../../app/data')
+
+const sendMessage = require('../mocks/modules/ffc-messaging')
 
 const { processMessage } = require('../../app/messaging/process-message')
 
@@ -113,16 +103,16 @@ describe('process statement', () => {
 
   test('sends 2 messages for publish and crm', async () => {
     await processMessage(message, receiver)
-    expect(mockSendMessage).toHaveBeenCalledTimes(2)
+    expect(sendMessage).toHaveBeenCalledTimes(2)
   })
 
   test('sends publish message with statement filename', async () => {
     await processMessage(message, receiver)
-    expect(mockSendMessage.mock.calls[0][0].body.filename).toBe(FILE_NAME)
+    expect(sendMessage.mock.calls[0][0].body.filename).toBe(FILE_NAME)
   })
 
   test('sends crm message with statement api link that contains filename', async () => {
     await processMessage(message, receiver)
-    expect(mockSendMessage.mock.calls[1][0].body.apiLink).toContain(FILE_NAME)
+    expect(sendMessage.mock.calls[1][0].body.apiLink).toContain(FILE_NAME)
   })
 })
