+++ test/mocks/modules/ffc-messaging.js
@@ -1,0 +1,13 @@
+const mockSendMessage = jest.fn()
+jest.mock('ffc-messaging', () => {
+  return {
+    MessageSender: jest.fn().mockImplementation(() => {
+      return {
+        sendMessage: mockSendMessage,
+        closeConnection: jest.fn()
+      }
+    })
+  }
+})
+
+module.exports = mockSendMessage
